Fix a bunch of bugs that had crept in because I was not running all the tests (or at least not correctly)
ElectricField class now does phase corrections to z=0 reference immediately on creation.  This simplifies doing calculations with E fields, and also fixes a phase shift bug that was not showing up in the fast tests but does appear in test_new_mie where we compare whole holograms.  

Change the E-field test in test theory to account for the phase shifted fields.  

Correct the handling of spherical coordinates.  It is not in fact correct to pass in 1xN Theta, 1xN Phi, and NxN R.  All three values need to be NxN since theta and phi both are computed from both x and y.  Now we pass mie_fields_sph a NxNx3 array containing thata, phi, and r.  

For now I have commented out the special cased computation for single sphere holograms.  That code does not work in spherical coordinates at the moment, we can reenable it when it does.  

Correct calc_intensity to just return abs(scat.x_comp)**2 + abs(scat.y_comp)**2.  Previously we were returning interfere_at_detector(scat, scat) which incorrectly (I think it is incorrect, it is at least different than it used to be) includes an interference term.