Move overlap checking into the spherecluster constructor, we never actually want a sphere cluster that has spheres overlapping.
Add a test of overlap checking

Add tests to multisphere to reject scatterers with very large particles or particle seperations.  It will not succeed in computing them, and they are usually typos anyway.  A python error is friendlier than a screenful of cryptic fortran output.

Add a test of multisphere bad input rejection.

Clean up a few tests that had gotten out of sync with some recent changes.