Prep work for allowing scatterers to be instantiated with parameters as arguments.  This means eliminating the checks and casts that assumed scatterer constructors were always supposed to be called with numbers.
Remove testing to confirm that arguments passed to scatterer constructors are numbers.  

Remove all the casts to complex of n for scatterers.  The number of parameters returned by Scatterer.parameter is variable for all scatterers now (it already was for SphereCluster and CoatedSphere).  

Remove RigidCluster, it was getting kind of out of date, and was the wrong way of doing things anyhow.  Its functionality will be recreated as make_scatterer generators.