in scatterpy.io.serializable
move __repr__ from Serializable to SerializeByConstructor, it was doing some things that needed introspection that we didn't want for Serializable (since they appear not to work correctly for mpfit objects).  mpfit objects now do not print, but that is probably ok, since they are very long and you usually don't want to see them.