Preliminary rework of Data into Marray
This changes names so that we are no longer referring to calculated
results as data.  Now we have Marray and Schema instead of data and DataTarget

Make the constructors of Marray (was Data) and subclasses a lot more
explicit about what they are doing with various metadata.

Make all Marray and Schema have a .shape attribute, this simplifies logic
in a number of places.  

Rework the inheritance tree of our storage objects to reduce diamond
problems.

This is a rough first cut, the tests still need to be reworked (which
probably means that there are still plenty of bugs hiding in the new
code), and the docs rewritten.