Refactor of marray
Change VectorImage to VectorGrid and clean it up for 3d work.  

move zeros_like out as a free function to better match how numpy does
it.  In the process I also got it working (hopefully) for all types of
marrays.  

Fix a few glitches in the __init__ chaining in the marray heirarchy.