Basic implementation of propagating e fields.
I have not done much to test the physical correctness of
this. Propagated fields do not strictly agree with either propagated
holograms or holograms calculated at the distance. Error, with respect
to direct calculation, appears to be lower for propagate field, then
convert to hologram than for propagate hologram, as I would expect, at
least in the simple cases I have tested so far.