Remove _version and the associated bzr call to set a minor version number.
Having an external system call as part of every import was slowing
down holopy imports by ~10% and also occasionally leading to weird
errors for users. I have cut that code, and replaced it with a minor
version number equal to the current bzr revision. That number was
mostly used for versioning on the docs, so just try to remember to
bump the minor version to the current bzr revision when updating the
docs. 

Minor version will probably be repurposed for bugfix releases after
2.0 is released anyway.