Delete some unused code in core.helpers.py to get it up to 100% coverage
Set warnings as errors in core.tests.common.py (which has the effect
of treating warnings as errors in most tests). This is better code hygiene.

Fix all the deprecations that change highlighted.

Switch another DDA test to ScattererByFunction instead of the old deprecated VoxelatedScatterer

Get rid of a bit of duplicated code across scattering tests.