Move most of the position handling code for Marrays into metadata.Positions.
This is the first in a series of commits to actually get scattering to
work for general Marrays (currently theories basically only work on grid Marrays).

That will in turn allow all of the subset code to move into fit,
allowing us to pull all of the subset stuff out of marray. Because of
that, this commit breaks some selection code and disables the
associated tests.