New implementation of random subset fiting!
 This implementation is now contained to a few lines in fit.py rather
than having tendrils through all of Marray. It takes advantage of the
newly finished ability to compute holograms of arbitrary Schema to
compute at the pixels instead of needing selection support explicitly
in marray.

This allows us to get rid of all of the selection code in holopy.core,
and removes the need for all of the selection tests.