regold subset tempering test
Hopefully this isn't masking an actual regression in the code, but enough has changed that I would not be surprised to see a change in code that uses randomness (ordering of rng calls probably changed meaning different things see different numbers, even though the seed is the same).

I did some basic investigation to see that inference wasn't way off, and fitting is working, so it is probably ok, but more testing is in order.