Decrease over-tight tolerances in test_lens.test_polarization_rotation_produces_small_changes_to_image
For this configuration of spheres, rotating the polarization seems to actually produce a ~2% change in the field intensity, and does not rotate the image (as checked on a full 2D grid of two spheres). The previous tolerances of 1e-3 were set when the polarization implementation of lens was incorrect, and were too tight.