mielensfunctions: Truncate series of al, bl to avoid nans:
The al, bl calculation can produce nan's if the maximum l value is made aggressively large, due to weirdness in the complex arithmetic standard. (The spherical Hankel functions can be (0 + 1j * inf) when l is large. But, due to the implementation of complex arithmetic in Python and numpy, 0 + 1j*inf gets cast as nan + 1j*inf. We then divide a non-infinite number by the spherical Hankel's nan + 1j*inf, and that gives nan's when it should give 0.) To avoid this, we truncate the series at a "reasonable" value of l while checking that no nans actually appear in the calculation. We do this by stopping the series if we get a nan, but checking that the previous term in the series is close to 0. (That comment is also in the code.) In addition to testing this, the test also tests if the test is correct, by seeing if numpy throws a complex arithmetic warning.